+++ client/src/components/create/message/lib/field/types/basicField.tsx
@@ -9,6 +9,7 @@ import { validateOptDataNum, validateOptDataBinary, validateOptDataStr } from '.
 import { hasProperty } from 'components/utils';
 import { FormattedField } from './Types';
 import { Buffer } from 'buffer';
+import SBECMARegexBtn from 'components/common/SBECMARegexBtn';
 
 // Interface
 interface BasicFieldProps {
@@ -28,6 +29,14 @@ const BasicField = (props: BasicFieldProps) => {
   const [idx, name, type, opts, comment] = def;
   const msgName = (parent ? [parent, name] : [name]).join('.');
   const [data, setData] = useState(value);
+
+  //used for ECMAScript regex checker button
+  const [isECMAScriptValid, setIsECMAScriptValid] = useState(false);
+  const [isValidating, setIsValidating] = useState(false);
+  const patternOnchg = (patternData: string) => {
+      setData(patternData);
+      optChange(name, patternData, arr);
+  }
 
   var optData: Record<string, any> = {};
   const schema = useAppSelector((state) => state.Util.selectedSchema) as SchemaJADN;
@@ -206,6 +215,47 @@ const BasicField = (props: BasicFieldProps) => {
     }
   }
 
+  if (optData.pattern)
+    return (
+      <div className='form-group'>
+        <div className='card'>
+          <div className={`card-header p-2 ${children ? 'd-flex justify-content-between' : ''}`}>
+            <div>
+              <p className='card-title m-0'>{`${name}${isOptional(def) ? '' : '*'}`}</p>
+              {comment ? <small className='card-subtitle form-text text-muted text-wrap'>{comment}</small> : ''}
+            </div>
+            {children}
+          </div>
+          <div className='card-body m-0 p-0'>
+            <input
+              type='text'
+              name={name}
+              value={data}
+              placeholder={optData.pattern ? optData.pattern : ''}
+              onChange={e => {
+                setData(e.target.value);
+              }}
+              onBlur={e => {
+                const errCheck = validateOptDataStr(config, optData, e.target.value);
+                setErrMsg(errCheck);
+                optChange(name, e.target.value, arr);
+              }}
+              style={{ borderColor: errMsg.length != 0 ? 'red' : '' }} /> 
+              <SBECMARegexBtn 
+                isECMAScriptValid={isECMAScriptValid}
+                setIsECMAScriptValid={setIsECMAScriptValid}
+                setIsValidating={setIsValidating}
+                patternData={optData.pattern}                         
+                onValidateClick={patternOnchg}
+              /> 
+    
+          </div>
+          {err}
+        </div>
+      </div>
+    )
+
+
   //DEFAULT: STRING 
   return (
     <div className='form-group'>
@@ -231,7 +281,7 @@ const BasicField = (props: BasicFieldProps) => {
               setErrMsg(errCheck);
               optChange(name, e.target.value, arr);
             }}
-            style={{ borderColor: errMsg.length != 0 ? 'red' : '' }} />
+            style={{ borderColor: errMsg.length != 0 ? 'red' : '' }} />            
         </div>
         {err}
       </div>